fix: avoid selectable labels for dynamic info
Make InfoRow labels not selectable by default. Allow an override of the attribute in the template (for Software tab for example).

Motivation:
GTK will place the blinking cursor on the first selectable label of a view/tab. Placing the cursor on dynamic info like the current GPU Core Clock will show the cursor jumping around as the number of digits change a lot (best seen in the OC tab).

As the selection of a label is discarded on update of a label (which happens multiple times a second for a lot of the OC and Thermals values) the cursor is just a visual distraction rather than a feature.